# SIGE-Desk/docs/openapi.yaml
{
  "openapi": "3.1.0",
  "info": {
    "title": "SIGE Desk API",
    "description": "API autenticada para gestão de demandas de tráfego.",
    "license": {
      "name": "Uso acadêmico"
    },
    "version": "v1"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Backend local"
    }
  ],
  "security": [
    {
      "sessionCookie": []
    }
  ],
  "paths": {
...
    "/api/v1/tickets": {
      "get": {
        "tags": [
          "ticket-controller"
        ],
        "operationId": "list_1",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "OPEN",
                "TRIAGE",
                "EXECUTION",
                "WAITING_FOR_CLIENT",
                "VALIDATION",
                "DONE",
                "REOPENED",
                "CANCELLED"
              ]
            }
          },
          {
            "name": "clientId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "campaignId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "demandTypeId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "priority",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "URGENT",
                "HIGH",
                "MEDIUM",
                "LOW"
              ]
            }
          },
          {
            "name": "assigneeId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "createdFrom",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "createdTo",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "overdue",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          },
          {
            "name": "page",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 0
            }
          },
          {
            "name": "size",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "default": 25
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TicketList"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Perfil ou CSRF não autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Recurso fora do escopo ou inexistente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Conflito de versão, transição ou cadastro",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "ticket-controller"
        ],
        "description": "Criação JSON ou multipart atômica. Multipart: request com Content-Type application/json, files para anexos gerais e field.<name> para cada campo FILE. Limite de 10 MB por arquivo e 11 MB por envio. Valores dos campos FILE são preenchidos pelo servidor.",
        "operationId": "createMultipart",
        "parameters": [
          {
            "name": "X-XSRF-TOKEN",
            "in": "header",
            "description": "Token obtido em GET /api/v1/auth/csrf",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "multipart/form-data": {
              "schema": {
                "type": "object",
                "properties": {
                  "request": {
                    "$ref": "#/components/schemas/CreateTicketRequest"
                  },
                  "files": {
                    "type": "array",
                    "items": {
                      "type": "string",
                      "format": "binary"
                    }
                  }
                },
                "required": [
                  "request"
                ]
              }
            },
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTicketRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TicketItem"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Perfil ou CSRF não autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Recurso fora do escopo ou inexistente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Conflito de versão, transição ou cadastro",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/v1/reports/tickets": {
      "get": {
        "tags": [
          "report-controller"
        ],
        "operationId": "report",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "OPEN",
                "TRIAGE",
                "EXECUTION",
                "WAITING_FOR_CLIENT",
                "VALIDATION",
                "DONE",
                "REOPENED",
                "CANCELLED"
              ]
            }
          },
          {
            "name": "clientId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "campaignId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "demandTypeId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "priority",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "URGENT",
                "HIGH",
                "MEDIUM",
                "LOW"
              ]
            }
          },
          {
            "name": "assigneeId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "createdFrom",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "createdTo",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "overdue",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/TicketReport"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Perfil ou CSRF não autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Recurso fora do escopo ou inexistente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Conflito de versão, transição ou cadastro",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/reports/tickets/export": {
      "get": {
        "tags": [
          "report-controller"
        ],
        "operationId": "export",
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "OPEN",
                "TRIAGE",
                "EXECUTION",
                "WAITING_FOR_CLIENT",
                "VALIDATION",
                "DONE",
                "REOPENED",
                "CANCELLED"
              ]
            }
          },
          {
            "name": "clientId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "campaignId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "demandTypeId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "priority",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "URGENT",
                "HIGH",
                "MEDIUM",
                "LOW"
              ]
            }
          },
          {
            "name": "assigneeId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "createdFrom",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "createdTo",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date"
            }
          },
          {
            "name": "overdue",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/csv": {
                "schema": {
                  "type": "string",
                  "format": "byte"
                }
              }
            }
          },
          "400": {
            "description": "Dados inválidos",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "403": {
            "description": "Perfil ou CSRF não autorizado",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Recurso fora do escopo ou inexistente",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "409": {
            "description": "Conflito de versão, transição ou cadastro",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "CreateTicketRequest": {
        "type": "object",
        "properties": {
          "clientId": {
            "type": "string",
            "minLength": 1
          },
          "campaignId": {
            "type": "string"
          },
          "pendingCampaign": {
            "type": "string",
            "maxLength": 255,
            "minLength": 0
          },
          "demandTypeId": {
            "type": "string",
            "minLength": 1
          },
          "channel": {
            "type": "string",
            "maxLength": 80,
            "minLength": 0
          },
          "subject": {
            "type": "string",
            "maxLength": 255,
            "minLength": 0
          },
          "description": {
            "type": "string",
            "minLength": 1
          },
          "urgency": {
            "type": "string",
            "enum": [
              "URGENT",
              "HIGH",
              "MEDIUM",
              "LOW"
            ]
          },
          "desiredDate": {
            "type": "string",
            "format": "date"
          },
          "requesterId": {
            "type": "string"
          },
          "metrics": {
            "$ref": "#/components/schemas/TicketMetrics"
          }
        },
        "required": [
          "channel",
          "clientId",
          "demandTypeId",
          "description",
          "subject",
          "urgency"
        ]
      },
      "TicketMetrics": {
        "type": "object",
        "properties": {
          "period": {
            "type": "string",
            "maxLength": 120,
            "minLength": 0
          },
          "impressions": {
            "type": "integer",
            "format": "int64"
          },
          "ctr": {
            "type": "number",
            "minimum": 0
          },
          "cpc": {
            "type": "number",
            "minimum": 0
          },
          "conversions": {
            "type": "integer",
            "format": "int64"
          },
          "cpa": {
            "type": "number",
            "minimum": 0
          },
          "roas": {
            "type": "number",
            "minimum": 0
          },
          "fields": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          }
        }
      },
      "TicketItem": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "clientId": {
            "type": "string"
          },
          "campaignId": {
            "type": [
              "string",
              "null"
            ]
          },
          "demandTypeId": {
            "type": "string"
          },
          "number": {
            "type": "string"
          },
          "subject": {
            "type": "string"
          },
          "clientName": {
            "type": "string"
          },
          "campaignName": {
            "type": [
              "string",
              "null"
            ]
          },
          "typeName": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "OPEN",
              "TRIAGE",
              "EXECUTION",
              "WAITING_FOR_CLIENT",
              "VALIDATION",
              "DONE",
              "REOPENED",
              "CANCELLED"
            ]
          },
          "priority": {
            "type": [
              "string",
              "null"
            ],
            "enum": [
              "URGENT",
              "HIGH",
              "MEDIUM",
              "LOW"
            ]
          },
          "urgency": {
            "type": "string",
            "enum": [
              "URGENT",
              "HIGH",
              "MEDIUM",
              "LOW"
            ]
          },
          "assigneeName": {
            "type": [
              "string",
              "null"
            ]
          },
          "updatedAt": {
            "type": "string",
            "format": "date-time"
          },
          "actions": {
            "type": "array",
            "items": {
              "type": "string"
            }
          },
          "version": {
            "type": "integer",
            "format": "int64"
          },
          "slaState": {
            "type": "string"
          },
          "overdue": {
            "type": "boolean"
          }
        }
      },
...
      "TicketList": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TicketItem"
            }
          },
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "size": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "ReportSummary": {
        "type": "object",
        "properties": {
          "activeCount": {
            "type": "integer",
            "format": "int64"
          },
          "completedCount": {
            "type": "integer",
            "format": "int64"
          },
          "classifiedCount": {
            "type": "integer",
            "format": "int64"
          },
          "slaCompliantCount": {
            "type": "integer",
            "format": "int64"
          },
          "overdueCount": {
            "type": "integer",
            "format": "int64"
          },
          "averageResolutionMinutes": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int64"
          }
        }
      },
      "TicketReport": {
        "type": "object",
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TicketItem"
            }
          },
          "total": {
            "type": "integer",
            "format": "int64"
          },
          "summary": {
            "$ref": "#/components/schemas/ReportSummary"
          }
        }
      },
...
      "ApiError": {
        "type": "object",
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "fields": {
            "type": "object",
            "additionalProperties": {
              "type": "string"
            }
          }
        }
      }
    },
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "name": "JSESSIONID",
        "in": "cookie"
      }
    }
  }
}